gpu: move the loop-body walks into their own unit
The walks that answer a question about a loop body without changing it --
which symbols it mentions, which locals it declares, which routines it
calls, whether it holds a nested section, a section of a dummy, an
implied-do, a function call, how many returns or nested scopes a callee
has -- are all read-only visitors over the same body, and the visitor is
their only caller.

Pure code motion: the definitions are unchanged, only their home is.
