+++ src/libasr/CMakeLists.txt
@@ -86,6 +86,7 @@ set(SRC
     pass/gpu_offload_designator.cpp
     pass/gpu_offload_preflight.cpp
     pass/gpu_offload_rewrite.cpp
+    pass/gpu_offload_collect.cpp
     pass/device_partition.cpp
     pass/device_launch_expand.cpp
     pass/gpu_memory_space.cpp
